Swift plugin (#630)

+++ src-tauri/src/service/mod.rs
@@ -574,7 +574,7 @@ impl From<InterfaceConfiguration> for proto::InterfaceConfig {
                 .map(ToString::to_string)
                 .collect::<Vec<_>>()
                 .join(","),
-            port: config.port,
+            port: u32::from(config.port),
             peers: config.peers.into_iter().map(Into::into).collect(),
         }
     }
@@ -591,7 +591,7 @@ impl From<proto::InterfaceConfig> for InterfaceConfiguration {
             name: config.name,
             prvkey: config.prvkey,
             addresses,
-            port: config.port,
+            port: config.port as u16,
             peers: config.peers.into_iter().map(Into::into).collect(),
             mtu: None,
         }
